// euc-lab-portal/src/app/models/Account.ts
export class Account {
    domain: string;
    badPwdCount: number;
    badPasswordTime: string;
    cn: string;
    displayName: string;
    distinguishedName: string;
    givenName: string;
    lockoutTime: any;
    logonCount: number;
    mail: string;
    objectGUID: any;
    sAMAccountName: string;
    sn: string;
    userAccountControl: number;
    userPrincipalName: string;
    firstname: string;
    lastname: string;
    username: string;
    upn: string;
    email: string;
    status: string;
    type: string;
    password: string;
}